# websrc/docs/swagger.yaml
basePath: /
definitions:
...
  models.BasicDataTask:
    properties:
      dummy:
        $ref: '#/definitions/models.GenFileParams'
      sourcePoint:
        $ref: '#/definitions/models.ProviderConfig'
      targetPoint:
        $ref: '#/definitions/models.ProviderConfig'
    type: object
  models.BasicResponse:
    properties:
      Error:
        type: string
      Result:
        type: string
    type: object
...
  models.MigrateTask:
    properties:
      dummy:
        $ref: '#/definitions/models.GenFileParams'
      operationId:
        type: string
      sourcePoint:
        $ref: '#/definitions/models.ProviderConfig'
      targetPoint:
        $ref: '#/definitions/models.ProviderConfig'
    type: object
...
  models.Schedule:
    properties:
      ScheduleID:
        type: string
      ScheduleName:
        type: string
      cron:
        type: string
      operationId:
        type: string
      startTime:
        type: string
      status:
        $ref: '#/definitions/models.Status'
      tag:
        items:
          type: string
        type: array
      tasks:
        items:
          $ref: '#/definitions/models.BasicDataTask'
        type: array
      tz:
        type: string
    type: object
  models.Status:
    enum:
    - active
    - inactive
    - pending
    - completed
    - failed
    type: string
    x-enum-varnames:
    - StatusActive
    - StatusInactive
    - StatusPending
    - StatusCompleted
    - StatusFailed
  models.Task:
    properties:
      operationId:
        type: string
      tag:
        items:
          type: string
        type: array
    type: object
info:
  contact:
    email: contact-to-cloud-barista@googlegroups.com
    name: API Support
    url: http://cloud-barista.github.io
  description: MC-Data-Manager REST API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: MC-Data-Manager REST API
  version: latest
paths:
...
  /migrate:
    get:
      description: Retrieve a list of all Tasks in the system.
      operationId: GetAllMigrateHandler
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved all Tasks
          schema:
            items:
              $ref: '#/definitions/models.Task'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.BasicResponse'
      summary: Get all Tasks
      tags:
      - '[Migrate]'
  /migrate/{id}:
    delete:
      description: Delete an existing Task using its ID.
      operationId: DeleteMigratekHandler
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully deleted the Task
          schema:
            $ref: '#/definitions/models.BasicResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/models.BasicResponse'
      summary: Delete a Task
      tags:
      - '[Migrate]'
    get:
      consumes:
      - application/json
      description: Get the details of a Task using its ID.
      operationId: GetMigrateHandler
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved a Task
          schema:
            $ref: '#/definitions/models.Task'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/models.BasicResponse'
      summary: Get a Task by ID
      tags:
      - '[Migrate]'
    put:
      consumes:
      - application/json
      description: Update the details of an existing Task using its ID.
      operationId: UpdateMigrateHandler
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      - description: Parameters required for updating a Task
        in: body
        name: RequestBody
        required: true
        schema:
          $ref: '#/definitions/models.Schedule'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated the Task
          schema:
            $ref: '#/definitions/models.BasicResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/models.BasicResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.BasicResponse'
      summary: Update an existing Task
      tags:
      - '[Migrate]'
...
  /migration/linux/aws:
    post:
      consumes:
      - application/json
      description: Migrate data stored in a Linux-based system to AWS S3.
      operationId: MigrationLinuxToS3PostHandler
      parameters:
      - description: Parameters required for migration
        in: body
        name: RequestBody
        required: true
        schema:
          $ref: '#/definitions/models.MigrateTask'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully migrated data
          schema:
            $ref: '#/definitions/models.BasicResponse'
        "400":
          description: Invalid Request
          schema:
            $ref: '#/definitions/models.BasicResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.BasicResponse'
      summary: Migrate data from Linux to AWS S3
      tags:
      - '[Migrate]'
  /migration/linux/gcp:
    post:
      consumes:
      - application/json
      description: Migrate data stored in a Linux-based system to GCP Cloud Storage.
      operationId: MigrationLinuxToGCPPostHandler
      parameters:
      - description: Parameters required for migration
        in: body
        name: RequestBody
        required: true
        schema:
          $ref: '#/definitions/models.MigrateTask'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully migrated data
          schema:
            $ref: '#/definitions/models.BasicResponse'
        "400":
          description: Invalid Request
          schema:
            $ref: '#/definitions/models.BasicResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.BasicResponse'
      summary: Migrate data from Linux to GCP Cloud Storage
      tags:
      - '[Migrate]'
  /migration/linux/ncp:
    post:
      consumes:
      - application/json
      description: Migrate data stored in a Linux-based system to NCP Object Storage.
      operationId: MigrationLinuxToNCPPostHandler
      parameters:
      - description: Parameters required for migration
        in: body
        name: RequestBody
        required: true
        schema:
          $ref: '#/definitions/models.MigrateTask'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully migrated data
          schema:
            $ref: '#/definitions/models.BasicResponse'
        "400":
          description: Invalid Request
          schema:
            $ref: '#/definitions/models.BasicResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.BasicResponse'
      summary: Migrate data from Linux to NCP Object Storage
      tags:
      - '[Migrate]'
...
  /migration/mysql:
    post:
      consumes:
      - application/json
      description: Migrate data from one MySQL database to another MySQL database.
      operationId: MigrationMySQLPostHandler
      parameters:
      - description: Parameters required for migration
        in: body
        name: RequestBody
        required: true
        schema:
          $ref: '#/definitions/models.MigrateTask'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully migrated data
          schema:
            $ref: '#/definitions/models.BasicResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.BasicResponse'
      summary: Migrate data from MySQL to MySQL
      tags:
      - '[Migrate]'
...
  /migration/nrdbms:
    post:
      consumes:
      - application/json
      description: Migrate data from NRDBMS to NRDBMS.
      operationId: MigrationNRDBMSPostHandler
      parameters:
      - description: Parameters required for migration
        in: body
        name: RequestBody
        required: true
        schema:
          $ref: '#/definitions/models.MigrateTask'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully migrated data
          schema:
            $ref: '#/definitions/models.BasicResponse'
        "400":
          description: Invalid Request
          schema:
            $ref: '#/definitions/models.BasicResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.BasicResponse'
      summary: Migrate data from NRDBMS to NRDBMS
      tags:
      - '[Migrate]'
  /migration/objectstorage:
    post:
      consumes:
      - application/json
      description: Migrate data from ObjectStorage to ObjectStorage.
      operationId: MigrationObjectstoragePostHandler
      parameters:
      - description: Parameters required for migration
        in: body
        name: RequestBody
        required: true
        schema:
          $ref: '#/definitions/models.MigrateTask'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully migrated data
          schema:
            $ref: '#/definitions/models.BasicResponse'
        "400":
          description: Invalid Request
          schema:
            $ref: '#/definitions/models.BasicResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.BasicResponse'
      summary: Migrate data from ObjectStorage to ObjectStorage
      tags:
      - '[Migrate]'
  /migration/rdbms:
    post:
      consumes:
      - application/json
      description: Migrate data from RDBMS to RDBMS.
      operationId: MigrationRDBMSPostHandler
      parameters:
      - description: Parameters required for migration
        in: body
        name: RequestBody
        required: true
        schema:
          $ref: '#/definitions/models.MigrateTask'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully migrated data
          schema:
            $ref: '#/definitions/models.BasicResponse'
        "400":
          description: Invalid Request
          schema:
            $ref: '#/definitions/models.BasicResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.BasicResponse'
      summary: Migrate data from RDBMS to RDBMS
      tags:
      - '[Migrate]'
  /migration/windows/aws:
    post:
      consumes:
      - application/json
      description: Migrate data stored in a Windows-based system to AWS S3.
      operationId: MigrationWindowsToS3PostHandler
      parameters:
      - description: Parameters required for migration
        in: body
        name: RequestBody
        required: true
        schema:
          $ref: '#/definitions/models.MigrateTask'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully migrated data
          schema:
            $ref: '#/definitions/models.BasicResponse'
        "400":
          description: Invalid Request
          schema:
            $ref: '#/definitions/models.BasicResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.BasicResponse'
      summary: Migrate data from Windows to AWS S3
      tags:
      - '[Migrate]'
  /migration/windows/gcp:
    post:
      consumes:
      - application/json
      description: Migrate data stored in a Windows-based system to GCP Cloud Storage.
      operationId: MigrationWindowsToGCPPostHandler
      parameters:
      - description: Parameters required for migration
        in: body
        name: RequestBody
        required: true
        schema:
          $ref: '#/definitions/models.MigrateTask'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully migrated data
          schema:
            $ref: '#/definitions/models.BasicResponse'
        "400":
          description: Invalid Request
          schema:
            $ref: '#/definitions/models.BasicResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.BasicResponse'
      summary: Migrate data from Windows to GCP Cloud Storage
      tags:
      - '[Migrate]'
...
swagger: "2.0"